Fix: Use moved modules correctly. #69
